0x13. C - More singly linked lists

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,4 +1,3 @@
-#include <string.h>
 #include <stdio.h>
 #include "lists.h"
 
@@ -11,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	unsigned int counter = 0;
+	size_t counter = 0;
 
 	if (h == NULL)
 		return (0);
